perf(frontend): store chunk-buffer timestamps as Float64Array, not BigInt64Array (L8)
The chunk buffer stored per-keyframe timestamps as a BigInt64Array. Every
fractional-index Hermite read (readBodyPositionInto, readBodyStateInto) ran
Number(timestamps[i0+1] - timestamps[i0]), and indexing a BigInt64Array
materializes a heap BigInt per element (the subtraction allocates a third).
Those reads run inside useFrame for every body in every consumer, so playback
churned on the order of 50 to 200 BigInt allocations per frame on a path the
render-loop rules require allocation-free.

Timestamps are Unix milliseconds. Float64 represents every integer exactly up
to 2^53 (about 285,000 years of ms); real epochs sit far below that (a
year-9999 epoch is about 2.5e14, roughly 36x under the ceiling), so float64 is
value-lossless here. Decoded timestamps and the Hermite dt are bit-identical to
before.

Swap the representation at every declare/alloc/read site and drop the now-dead
BigInt()/Number() round-trips: the chunk buffer, the binary-chunk decoder, the
true-track rebuild, the two payload interfaces, and the DevPanel memory
readout. The wire format is unchanged: the int64 start timestamp is still read
with getBigInt64 then wrapped in Number, exactly as before. No backend, DTO,
OpenAPI, or wire-version change.

Float64Array and BigInt64Array are mutually incompatible at assignment and
.set(), so any missed site is a tsc error, not a silent bug; the migration is
all-or-nothing by construction. A new guard test pins both the representation
(timestamps instanceof Float64Array) and the exactness bound (a year-9999
decode is exact). Four test fixtures beyond the originally scoped set also
carried BigInt64Array timestamps and were migrated.

Gates: tsc (zero new errors, one stale BigInt-literal error removed), eslint,
358 vitest tests, and next build all green.

+++ frontend/src/app/store/middleware/groundTruthMiddleware.test.ts
@@ -28,8 +28,8 @@ const T0 = 1_750_000_000_000;
 const N = 9_000; // keyframes in the buffer (> LOOKAHEAD_KEYFRAMES = 4000)
 
 function chunkPayload(n: number, startMs: number) {
-  const timestamps = new BigInt64Array(n);
-  for (let i = 0; i < n; i++) timestamps[i] = BigInt(startMs + i * HOUR_MS);
+  const timestamps = new Float64Array(n);
+  for (let i = 0; i < n; i++) timestamps[i] = startMs + i * HOUR_MS;
   return {
     bodyNames: ["Mars"],
     bodyCount: 1,
